+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 - demos: Fixed decimal overflow in `demo_uniswap` project.
 - evm.node: Fixed incorrect log request parameters.
+- cli: Added warning for timescaledb-ha users
 
 ## [7.0.0] - 2023-09-25